chore: move hybridObjects to `cpp/hybridObjects` folder
