+++ package/cpp/hybridObjects/HybridNitroSQLite.hpp
