+++ package/cpp/hybridObjects/HybridNitroSQLiteQueryResult.cpp
