+++ package/cpp/hybridObjects/HybridNitroSQLiteQueryResult.hpp
